improve: Remove superfluous _setRoleAdmin call for UPDATER_ROLE (#19)
* improve: Remove superfluous _setRoleAdmin call for UPDATER_ROLE

The SP1Helios contract inherits the AccessControlEnumerable contract to enable role-based access control mechanisms. According to the comments, the UPDATER_ROLE should be admin-less, making the updaters set immutable once deployed. As mentioned in the AccessControl contract, by default, the admin role for all roles is DEFAULT_ADMIN_ROLE, which means that only accounts with this role will be able to grant or revoke other roles.

However, the DEFAULT_ADMIN_ROLE's value is set to bytes32(0), making the call in line 141, where the UPDATER_ROLE's admin is being set to bytes32(0), superfluous. Moreover, the comment in the test in line 70 in SP1Helios.t.sol suggests that UPDATER_ROLE should be its own admin, which contradicts the previously mentioned comment in SP1Helios, that the UPDATER_ROLE should be admin less. In addition, the subsequent check in line 71 is only checking that the initialUpdater has the UPDATER_ROLE, and not that the initialUpdater is the admin of the UPDATER_ROLE.

Consider removing the _setRoleAdmin call in SP1Helios to ensure that UPDATER_ROLE has no admin. Since no address is assigned to the DEFAULT_ADMIN_ROLE, the UPDATER_ROLE's admin will remain DEFAULT_ADMIN_ROLE, with no address assigned to this role. Additionally, consider fixing the comment in the test to align with the documentation in the SP1Helios contract.

* Update SP1Helios.t.sol

* Update SP1Helios.t.sol